+++ apps-intern/ungdomsytelse-veileder/mock/data/skjermetDeltaker.ts
@@ -31,6 +31,8 @@ const deltakelse: DeltakelseDto = {
     tilOgMed: '2026-01-01',
     erSlettet: false,
     harOpphørsvedtak: false,
+    harUtvidetKvote: false,
+    kvoteMaksDato: '2026-01-01',
     søktTidspunkt: '2025-05-31T03:58:29.015999Z',
 };